Implement malloc system call + stdlib wrapper
I implemented a malloc system call which can be used by user processes
to request memory from the kernel. The memory is allocated from the
kernel's heap, and is directly mapped into the page tables of the user
process.

I also added a wrapper for this system call to the stdlib.