Implement free system call + stdlib wrapper
I implemented a free system call which can be used by user processes to
free memory that the process allocated through malloc. Once memory is
freed, the page tables for the process which allow it to write to that
address are reset.

I also added a wrapper for this system call to the stdlib, and updated
my stdlib test program to use this wrapper.