Implemented itoa in stdlib
Created the itoa function to convert base 10 integers into their string
representation.

I also removed fpic from our build options for the test_stdlib program
because our loader does not know how to handle global offset tables.
This option is not required because the test_stdlib executable is not a
shared library - it is loaded into a known place in memory.